# Backend_Jobaile/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobaile API",
        "description": "Dokumentasi API untuk aplikasi Jobaile (versi pertama)",
        "contact": {
            "name": "Jobaile Dev Team",
            "email": "support@jobaile.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/review/{id_reviewed}": {
            "get": {
                "tags": [
                    "RatingReview"
                ],
                "summary": "Lihat semua rating yang diterima user tertentu",
                "operationId": "e2a1137eb297d3fb568a40d4dab2f40c",
                "parameters": [
                    {
                        "name": "id_reviewed",
                        "in": "path",
                        "description": "ID user yang ingin dilihat ratingnya",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USR123"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter rating (opsional, 1-5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3,
                                4,
                                5
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data rating",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data rating."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "reviewer_nama": {
                                                        "type": "string",
                                                        "example": "Andi Budi"
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "ulasan": {
                                                        "type": "string",
                                                        "example": "Sangat baik"
                                                    },
                                                    "tanggal_rating": {
                                                        "type": "string",
                                                        "example": "2025-06-23 10:12"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "worker"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Belum login"
                    },
                    "403": {
                        "description": "Email belum diverifikasi"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RatingReview"
                ],
                "summary": "Beri atau update rating",
                "operationId": "e10f72e934f69a682e5ecc872bd85471",
                "parameters": [
                    {
                        "name": "id_reviewed",
                        "in": "path",
                        "description": "ID user yang diberi rating",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USR123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "ulasan": {
                                        "type": "string",
                                        "example": "Kerja bagus dan profesional."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil menyimpan rating",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating dan/atau ulasan berhasil disimpan."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Tidak bisa memberi rating ke diri sendiri atau sesama role"
                    },
                    "401": {
                        "description": "Belum terautentikasi"
                    },
                    "403": {
                        "description": "Email belum diverifikasi"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "RatingReview"
                ],
                "summary": "Update rating jika sudah ada",
                "operationId": "b52544768ef1e0872ee45345993e98e0",
                "parameters": [
                    {
                        "name": "id_reviewed",
                        "in": "path",
                        "description": "ID user yang diberi rating",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USR123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "ulasan": {
                                        "type": "string",
                                        "example": "Cukup membantu, perlu ditingkatkan."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating berhasil diperbarui"
                    },
                    "400": {
                        "description": "Tidak valid memberi rating"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "403": {
                        "description": "Email belum diverifikasi"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/review": {
            "get": {
                "tags": [
                    "RatingReview"
                ],
                "summary": "Lihat semua rating yang saya terima",
                "operationId": "33c9b89c79f9daa88b926c24bd2e6179",
                "parameters": [
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter berdasarkan rating (1-5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3,
                                4,
                                5
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data rating saya",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data rating Anda."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "reviewer_nama": {
                                                        "type": "string",
                                                        "example": "Dewi Kartika"
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "ulasan": {
                                                        "type": "string",
                                                        "example": "Top performer"
                                                    },
                                                    "tanggal_rating": {
                                                        "type": "string",
                                                        "example": "2025-06-23 09:01"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "recruiter"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Belum login"
                    },
                    "403": {
                        "description": "Email belum diverifikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}